# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches/{matchId}/substitutions:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    get:
      operationId: listMatchSubstitutions
      tags: [matches]
      summary: List all substitutions for a match (across all its sets), for rebuilding lineups
      responses:
        "200":
          description: List of substitutions, ordered by set then score
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Substitution"
  /sets/{setId}/substitutions:
    parameters:
      - $ref: "#/components/parameters/SetId"
    post:
      operationId: createSubstitution
      tags: [matches]
      summary: Record one substitution (regular sub or libero in/out) within a set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSubstitution"
      responses:
        "201":
          description: Created substitution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Substitution"

  # 刪掉一筆換人紀錄（前端「復原」退掉上一個換人動作時用，見 issue #41）。
  # 換人是 append-only 的，undo 靠 hard-delete 這一筆 row（不寫補償 row）——因為換人的
  # 記錄時機是「比分快照」而非明確時間戳，補償 row 的比分較低、reload 重放時會排在前面
  # 空轉、幽靈紀錄反而留著，所以只能真的把 row 刪掉（詳見 issue #41 討論）。
  /substitutions/{substitutionId}:
    parameters:
      - $ref: "#/components/parameters/SubstitutionId"
    delete:
      operationId: deleteSubstitution
      tags: [matches]
      summary: Delete a substitution (e.g. undo the last substitution action)
      responses:
        "204":
          description: Deleted

  # 暫停紀錄（issue #44）。整組路由刻意跟 substitutions 平行：讀（bulk）掛在
  # /matches/{matchId}/timeouts、寫掛在 /sets/{setId}/timeouts、undo 走 DELETE /timeouts/{id}。
  /matches/{matchId}/timeouts:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    get:
      operationId: listMatchTimeouts
      tags: [matches]
      summary: List all timeouts for a match (across all its sets), for rebuilding the scoresheet
      responses:
        "200":
          description: List of timeouts, ordered by set then score
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Timeout"
  /sets/{setId}/timeouts:
    parameters:
      - $ref: "#/components/parameters/SetId"
    post:
      operationId: createTimeout
      tags: [matches]
      summary: Record one timeout within a set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTimeout"
      responses:
        "201":
          description: Created timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timeout"

  # 刪掉一筆暫停紀錄（前端「復原」退掉上一個暫停動作時用）。跟 substitutions 一樣是 append-only、
  # 記錄時機是比分快照，所以 undo 只能 hard-delete 這一筆 row（見 substitutions 的同段說明）。
  /timeouts/{timeoutId}:
    parameters:
      - $ref: "#/components/parameters/TimeoutId"
    delete:
      operationId: deleteTimeout
      tags: [matches]
      summary: Delete a timeout (e.g. undo the last timeout action)
      responses:
        "204":
          description: Deleted

  # 一場比賽每一局的起始先發（我方六個號位 1~6 各站哪個球員；自由球員不列入，見
  # lib/db/src/schema/lineups.ts）。讀（bulk）掛在 /matches/:matchId/lineups 底下
  # （先驗 match 擁有權），道理跟 /matches/:matchId/substitutions 一樣：前端進計分表
  # 重建時一支請求把整場所有局的先發拿回來，不用對每個 set 各發一次。
...
  /tactics:
    get:
      operationId: listTactics
      tags: [tactics]
      summary: List all saved tactics for the current user
      parameters:
        # 依比賽過濾戰術庫（#119）。帶 matchId 就只回這場的戰術，戰術面板才不會跨場汙染。
        # 不帶就回全部（保留舊行為）。
        - name: matchId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of tactics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tactic"
    post:
      operationId: createTactic
      tags: [tactics]
      summary: Save a new tactic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTactic"
      responses:
        "201":
          description: Created tactic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tactic"
  /tactics/{tacticId}:
    parameters:
      - $ref: "#/components/parameters/TacticId"
    get:
      operationId: getTactic
      tags: [tactics]
      summary: Get a single tactic by id
      responses:
        "200":
          description: The tactic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tactic"
    put:
      operationId: updateTactic
      tags: [tactics]
      summary: Overwrite an existing tactic (name and/or data)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTactic"
      responses:
        "200":
          description: Updated tactic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tactic"
    delete:
      operationId: deleteTactic
      tags: [tactics]
      summary: Delete a tactic
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    MatchId:
      name: matchId
      in: path
      required: true
      schema:
        type: integer
...
    SetId:
      name: setId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SubstitutionId:
      name: substitutionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TimeoutId:
      name: timeoutId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TacticId:
      name: tacticId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    SubstitutionKind:
      type: string
      enum: [regular, libero]
    Substitution:
      type: object
      properties:
        # id/setId 從整數改成 uuid（#64 PR1），理由同 MatchSet.id。
        id:
          type: string
          format: uuid
        setId:
          type: string
          format: uuid
        # 換人發生的「時機」，存的是這一局當下的比分快照（而非 rallyId）——
        # 理由跟 Rally 的 homeScore/awayScore 一樣，見 lib/db/src/schema/substitutions.ts。
        homeScore:
          type: integer
        awayScore:
          type: integer
        # 上場、下場的球員；nullable，理由見 substitutions.ts：即使不知道是「誰」，
        # 「這一局發生過一次換人」本身仍有統計意義，球員被刪除時不該連帶刪掉這筆紀錄。
        playerInId:
          type: string
          format: uuid
          nullable: true
        playerOutId:
          type: string
          format: uuid
          nullable: true
        kind:
          $ref: "#/components/schemas/SubstitutionKind"
        # 伺服器指派的插入順序流水號，只給「同一分內排序」用，不是給前端識別身分用的。
        # id 改成 uuid 之後不再帶時序資訊，不能拿來當 orderBy 的 tiebreaker（見
        # lib/db/src/schema/substitutions.ts、artifacts/api-server/src/routes/substitutions.ts
        # 的 orderBy 註解），所以另外開一個欄位。這是 row 的真實欄位（getTableColumns 會回傳
        # 它），所以要宣告在回應 schema 裡，不然跟實際回應對不上；但 NewSubstitution 不能有
        # 這個欄位——它是 server-assigned，前端不該送。
        seq:
          type: integer
      required: [id, setId, homeScore, awayScore, playerInId, playerOutId, kind, seq]
    NewSubstitution:
      type: object
      properties:
        # 選填的 client-mintable 主鍵，理由同 NewSet.id（#64 PR2）。
        id:
          type: string
          format: uuid
        homeScore:
          type: integer
        awayScore:
          type: integer
        playerInId:
          type: string
          format: uuid
          nullable: true
        playerOutId:
          type: string
          format: uuid
          nullable: true
        kind:
          $ref: "#/components/schemas/SubstitutionKind"
      required: [homeScore, awayScore, kind]

    # 暫停是哪一方叫的（issue #44）。跟 RallyWinner 一樣用 home/away（前端 us/opponent 在
    # mapping 層翻譯）。獨立一個 enum、不共用 RallyWinner，對齊後端各表各自定 pgEnum 的慣例。
    TimeoutSide:
      type: string
      enum: [home, away]
    Timeout:
      type: object
      properties:
        # id/setId 從整數改成 uuid（#64 PR1），理由同 MatchSet.id。
        id:
          type: string
          format: uuid
        setId:
          type: string
          format: uuid
        # 暫停發生的「時機」＝這一局當下的比分快照（而非 rallyId），理由見 lib/db/src/schema/timeouts.ts。
        homeScore:
          type: integer
        awayScore:
          type: integer
        side:
          $ref: "#/components/schemas/TimeoutSide"
        # 伺服器指派的插入順序流水號，理由跟 Substitution.seq 完全一樣——id 改成隨機 uuid
        # 後不能再當排序 tiebreaker，另開一個單純遞增的欄位給 orderBy 用。
        seq:
          type: integer
      required: [id, setId, homeScore, awayScore, side, seq]
    NewTimeout:
      type: object
      properties:
        # 選填的 client-mintable 主鍵，理由同 NewSet.id（#64 PR2）。
        id:
          type: string
          format: uuid
        homeScore:
          type: integer
        awayScore:
          type: integer
        side:
          $ref: "#/components/schemas/TimeoutSide"
      required: [homeScore, awayScore, side]

    # 一局的起始先發：我方六個號位（1~6）各站哪個球員。詳見 lib/db/src/schema/lineups.ts。
    # 自由球員（L）不出現在這六個號位——前端邏輯是 L 從場邊出發、換人上場，開局先發只有這 6 人。
    # 「六人不得重複」DB 層表達不了，交給前端/route 檢查（跟 schema 註解一致）。
...
    Tactic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        # 這份戰術所屬的比賽 id（#119）。matches.id 是 serial 整數，所以這裡是 integer 不是 uuid。
        # null 代表 #119 之前存的舊戰術（沒有歸屬某場）。
        matchId:
          type: integer
          nullable: true
        name:
          type: string
        # TacticsState 的完整快照，以 jsonb 存在 DB，前端直接解析成 TacticsState 型別。
        data:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, data, createdAt, updatedAt]
    NewTactic:
      type: object
      properties:
        # 存檔時帶上當前比賽 id，讓這份戰術歸屬到這場（#119）。optional/nullable：不帶就是全域戰術，
        # 保留舊行為的相容性，但正常路徑前端都會帶。
        matchId:
          type: integer
          nullable: true
        name:
          type: string
        data:
          type: object
          additionalProperties: true
      required: [name, data]
    UpdateTactic:
      type: object
      properties:
        name:
          type: string
        data:
          type: object
          additionalProperties: true
...